+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * Automatically download the packages referenced in a given `Cargo.lock` file.
   * Parse the metadata of `Cargo.toml` files.
 * Allow setting the log level for the CLI.
+* Add option to prefer/download source distributions from PyPI instead of wheels.
 
 # Version 0.9.0 - 2024-03-23